[mail] Tweak timer and add timeout
This timer should activate about a minute after a successful boot and
then a minute after the last *finished* run.

However if something happens and offlineimap.service is unable to finish
a sync for 120 seconds kill it and wait another minute. The service will
not start if there is *not* configured network connection, but in case
the latter go away while syncing is in progress it will basically hang
for good (still not sure why offlineimap is not handling this).