bash: reuse as much as possible from ./sh
Spent the last year or so using bash as my shell. I would've probably
stayed here for good, given that bash is still to be considered as
default, if it had proper right prompt support.

There is a flaky emulation of such in this config, but it tends to annoy
me often enough for me to ditch it.