mail: rm -rf
mutt is great and I really wanted to keep reading mails from the
terminal. However, the rest of the world does not agree. Most of the
emails I receive are with rich content and lacking or incomplete plain
version, and I'm not taking part of a project where receiving pull
requests over email is a thing, anyways.

KMail is what I use nowadays, but the configuration there involves
akonadi and storing stuff in relational database.